requirements.txt
APScheduler==3.6.3
backcall==0.2.0
certifi==2020.12.5
charset-normalizer==2.0.4
decorator==5.0.9
idna==3.2
ipdb==0.13.9
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
python-decouple==3.4
python-dotenv==0.15.0
python-telegram-bot==13.3
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.3
wcwidth==0.2.5
